+++ src/3rdParty/salomesmesh/inc/SMESH_File.hxx
@@ -1,4 +1,4 @@
-// Copyright (C) 2007-2015  CEA/DEN, EDF R&D, OPEN CASCADE
+// Copyright (C) 2007-2016  CEA/DEN, EDF R&D, OPEN CASCADE
 //
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Lesser General Public
@@ -109,7 +109,7 @@ public:
 private:
 
   std::string _name; //!< file name
-  int         _size; //!< file size
+  long        _size; //!< file size
   std::string _error;
 #ifdef WIN32
   HANDLE      _file, _mapObj;
